+++ fivetran/tests/e2e/resource_private_link_e2e_test.go
@@ -25,7 +25,7 @@ var privateLinkResourceConfig = `
             	}`
 
 func TestResourcePrivateLinkE2E(t *testing.T) {
-	//t.Skip("Private links have a strict limit on the number of entities created. This test should only be used for intermediate tests when changes are made directly to Private links.")
+	t.Skip("Private links have a strict limit on the number of entities created. This test should only be used for intermediate tests when changes are made directly to Private links.")
 	suffix := strconv.Itoa(seededRand.Int())
 	privateLinkName := suffix
 	privateLinkCfgValue := "privatelink_" + suffix